+++ src/rules/media-query-list-comma-newline-after/__tests__/index.js
@@ -29,35 +29,41 @@ testRule(rule, {
   }, {
     code: "@media screen and (color)\r\n,\r\n\t\t\tprojection and (color) {}",
     description: "indentation after the CRLF after the comma",
+  }, {
+    code: "@non-media screen and (color),projection and (color) {}",
+    description: "ignore at-rules contain media in name",
+  }, {
+    code: "@media-non screen and (color),projection and (color) {}",
+    description: "ignore at-rules contain media in name",
   } ],
 
   reject: [ {
-    code: "@media screen and (color),projection and (color)",
+    code: "@media screen and (color),projection and (color) {}",
     message: messages.expectedAfter(),
     line: 1,
     column: 26,
   }, {
-    code: "@mEdIa screen and (color),projection and (color)",
+    code: "@mEdIa screen and (color),projection and (color) {}",
     message: messages.expectedAfter(),
     line: 1,
     column: 26,
   }, {
-    code: "@MEDIA screen and (color),projection and (color)",
+    code: "@MEDIA screen and (color),projection and (color) {}",
     message: messages.expectedAfter(),
     line: 1,
     column: 26,
   }, {
-    code: "@media screen and (color), projection and (color)",
+    code: "@media screen and (color), projection and (color) {}",
     message: messages.expectedAfter(),
     line: 1,
     column: 26,
   }, {
-    code: "@media screen and (color),  projection and (color)",
+    code: "@media screen and (color),  projection and (color) {}",
     message: messages.expectedAfter(),
     line: 1,
     column: 26,
   }, {
-    code: "@media screen and (color),\tprojection and (color)",
+    code: "@media screen and (color),\tprojection and (color) {}",
     message: messages.expectedAfter(),
     line: 1,
     column: 26,
@@ -92,6 +98,12 @@ testRule(rule, {
   }, {
     code: "@media screen and (color),projection and (color) {\r\n}",
     description: "ignore single line list, multi-line block and CRLF",
+  }, {
+    code: "@non-media screen and (color),projection and (color),\nprint {}",
+    description: "ignore at-rules contain media in name",
+  }, {
+    code: "@media-non screen and (color),projection and (color),\nprint {}",
+    description: "ignore at-rules contain media in name",
   } ],
 
   reject: [ {
@@ -151,6 +163,12 @@ testRule(rule, {
   }, {
     code: "@media screen and (color), projection and (color) {\n}",
     description: "ignore single line list, multi-line block",
+  }, {
+    code: "@non-media screen and (color) ,projection and (color),\nprint {}",
+    description: "ignore at-rules contain media in name",
+  }, {
+    code: "@media-non screen and (color) ,projection and (color),\nprint {}",
+    description: "ignore at-rules contain media in name",
   } ],
 
   reject: [ {