+++ src/rules/media-query-list-comma-newline-before/__tests__/index.js
@@ -31,30 +31,36 @@ testRule(rule, {
     description: "indentation after the CRLF before the comma",
   }, {
     code: "@media screen and (color)\n\n, projection and (color)",
+  }, {
+    code: "@non-media screen and (color), projection and (color) {}",
+    description: "ignore at-rules contain media in name",
+  }, {
+    code: "@media-non screen and (color), projection and (color) {}",
+    description: "ignore at-rules contain media in name",
   } ],
 
   reject: [ {
-    code: "@media screen and (color), projection and (color)",
+    code: "@media screen and (color), projection and (color) {}",
     message: messages.expectedBefore(),
     line: 1,
     column: 26,
   }, {
-    code: "@mEdIa screen and (color), projection and (color)",
+    code: "@mEdIa screen and (color), projection and (color) {}",
     message: messages.expectedBefore(),
     line: 1,
     column: 26,
   }, {
-    code: "@MEDIA screen and (color), projection and (color)",
+    code: "@MEDIA screen and (color), projection and (color) {}",
     message: messages.expectedBefore(),
     line: 1,
     column: 26,
   }, {
-    code: "@media screen and (color)  , projection and (color)",
+    code: "@media screen and (color)  , projection and (color) {}",
     message: messages.expectedBefore(),
     line: 1,
     column: 28,
   }, {
-    code: "@media screen and (color)\t, projection and (color)",
+    code: "@media screen and (color)\t, projection and (color) {}",
     message: messages.expectedBefore(),
     line: 1,
     column: 27,
@@ -86,6 +92,12 @@ testRule(rule, {
   }, {
     code: "@media screen and (color),projection and (color) {\n}",
     description: "ignore single line list, multi-line block",
+  }, {
+    code: "@non-media screen and (color),projection and (color)\n, print {}",
+    description: "ignore at-rules contain media in name",
+  }, {
+    code: "@media-non screen and (color),projection and (color)\n, print {}",
+    description: "ignore at-rules contain media in name",
   } ],
 
   reject: [ {
@@ -142,6 +154,12 @@ testRule(rule, {
   }, {
     code: "@media screen and (color) ,projection and (color) {\n}",
     description: "ignore single line list, multi-line block",
+  }, {
+    code: "@non-media screen and (color) ,projection and (color),\nprint {}",
+    description: "ignore at-rules contain media in name",
+  }, {
+    code: "@media-non screen and (color) ,projection and (color),\nprint {}",
+    description: "ignore at-rules contain media in name",
   } ],
 
   reject: [ {